+++ lib/music/music_manager.rb
@@ -10,7 +10,7 @@ class MusicManager
   def initialize
     @list_genre = []
     @list_album = []
-		load_json
+    load_json
   end
 
   def validated_boolean(option)
@@ -24,12 +24,12 @@ class MusicManager
     end
   end
 
-	def genre_menu
+  def genre_menu
     puts 'Add album genre:'
     genre_name = gets.chomp
     genre = Genre.new(genre_name)
     @list_genre << genre
-		genre
+    genre
   end
 
   def album_menu
@@ -44,8 +44,8 @@ class MusicManager
     spotify = gets.chomp.upcase
     check_spoti = validated_boolean(spotify)
     album = MusicAlbum.new(album_name, date, check_ar, check_spoti)
-		genre = genre_menu
-		genre.add_item(album)
+    genre = genre_menu
+    genre.add_item(album)
     puts 'Album created succesfully'
     @list_album << album
   end
@@ -62,27 +62,28 @@ class MusicManager
     end
   end
 
-	def write_json
-		arr = []
-		@list_album.each do |album|
-			arr.push({name: album.name, genre: album.genre.name, published: album.publish_date,
-				archived: album.archived, spotify: album.on_spotify, id: album.id})
-		end
-		FileUtils.touch('music.json') unless File.exist?('music.json')
-		File.write('music.json', JSON.pretty_generate(arr))
-	end
+  def write_json
+    arr = []
+    @list_album.each do |album|
+      arr.push({ name: album.name, genre: album.genre.name, published: album.publish_date,
+                 archived: album.archived, spotify: album.on_spotify, id: album.id })
+    end
+    FileUtils.touch('music.json') unless File.exist?('music.json')
+    File.write('music.json', JSON.pretty_generate(arr))
+  end
 
-	def load_json
-		return unless File.exist?('music.json')
-		file = JSON.parse(File.read('music.json'))
-		file.each do |data|
-			album = MusicAlbum.new(data['name'], data['published'], data['archived'],
-				data['spotify'], data['id'])
-				genre = Genre.new(data['genre'])
-				album.genre = genre
-				genre.add_item(album)
-				@list_album.push(album)
-				@list_genre.push(genre)
-		end
-	end
+  def load_json
+    return unless File.exist?('music.json')
+
+    file = JSON.parse(File.read('music.json'))
+    file.each do |data|
+      album = MusicAlbum.new(data['name'], data['published'], data['archived'],
+                             data['spotify'], data['id'])
+      genre = Genre.new(data['genre'])
+      album.genre = genre
+      genre.add_item(album)
+      @list_album.push(album)
+      @list_genre.push(genre)
+    end
+  end
 end
