+++ spec/spec_helper.rb
@@ -16,7 +16,6 @@
 require_relative '../lib/books/book'
 require_relative '../lib/books/label'
 
-
 RSpec.configure do |config|
   # rspec-expectations config goes here. You can use an alternate
   # assertion/expectation library such as wrong or the stdlib/minitest
